v.1.0
The first working prototype of the parser. A server was created on the Django framework. The parser sends a request to the marketplace, and then displays it on the localhost. If you want to take a look at the parser code, then you need the pages / engine.py file. In the next versions, it is necessary to work out errors when scratching information from the source, for example, when the query "Пистолет" appears, results containing information about the remaining quantity appear, which the parser cannot process at the moment. As for the rest of the implementation, it's necessary to: correctly display the received elements, disable the display of the web driver, configure the pages to work, create a start page.